XP Level is now saved with Double precision
Progression now takes into account the exact XP level instead of the Int level, making progress gain curve smoother (and usually lower xp gain)
Progression is no longer rounded down

+++ BreakBaloon Watch Extension/ExtensionDelegate.swift
@@ -66,12 +66,12 @@ class ExtensionDelegate: NSObject, WKExtensionDelegate, WCSessionDelegate {
     
     func session(_ session: WCSession, didReceiveUserInfo userInfo: [String: Any] = [:]) {
         print("receive raw: \(userInfo)")
-        if let exp = userInfo["exp"] as? Int {
-            if UserDefaults.standard.integer(forKey: "exp") <= exp {
+        if let exp = userInfo["exp"] as? Double {
+            if UserDefaults.standard.double(forKey: "exp") <= exp {
                 UserDefaults.standard.set(exp, forKey: "exp")
                 print("Received XP:", exp)
             } else {
-                session.transferUserInfo(["exp": UserDefaults.standard.integer(forKey: "exp")])
+                session.transferUserInfo(["exp": UserDefaults.standard.double(forKey: "exp")])
             }
         }
         if let animate = userInfo["extension.animation.enabled"] as? Bool {
